Update product model and UI navigation logic
- Update `Product` model to use `int` for price fields and add multiple constructors for flexibility.
- Refactor `ProductCardAdapter` to support dynamic layout resource injection.
- Enhance `MainActivity` routing logic to handle "Get Started" and targeted navigation.
- Improve `ProductDetail` and `Homepage` data binding by converting numeric prices to formatted strings for the UI.
- Update `NavBarItem` and `NavBarIcon` styles to use a new color selector and improved layout weights.
- Include header and navigation bar in `AboutUsActivity` layout.
- Rename `app_name` to "TEATRACK" in string resources.
- Adjust Gradle JVM arguments and disable parallel builds in `gradle.properties`.

+++ app/src/main/java/com/teatrack_mcd_253eie502802_group02/data/FirebaseProductRepository.java
@@ -105,10 +105,10 @@ public class FirebaseProductRepository {
         }
 
         // Firebase schema uses "price" for size M.
-        product.setPrice(readLong(snapshot, "price"));
-        product.setPriceL(readLong(snapshot, "priceL"));
-        product.setVipPriceM(readLong(snapshot, "vipPriceM"));
-        product.setVipPriceL(readLong(snapshot, "vipPriceL"));
+        product.setPrice((int) readLong(snapshot, "price"));
+        product.setPriceL((int) readLong(snapshot, "priceL"));
+        product.setVipPriceM((int) readLong(snapshot, "vipPriceM"));
+        product.setVipPriceL((int) readLong(snapshot, "vipPriceL"));
         product.setImages(readImages(snapshot));
         return product;
     }
